pyproject.toml
[tool.poetry]
name = "plant-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
tqdm = "^4.66.1"
unstructured = {extras = ["pdf"], version = "^0.12.2"}
langchain-experimental = "^0.0.49"
python-dotenv = "^1.0.0"
openai = "^1.9.0"
tiktoken = "^0.5.2"
langchain-openai = "^0.0.3"
psycopg = "^3.1.17"
pgvector = "^0.2.4"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
